pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "needicons"
version = "0.1.1"
description = "AI-powered icon pack generator with professional post-processing"
requires-python = ">=3.10"
license = "AGPL-3.0-or-later"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
    "pillow>=10.0",
    "opencv-python-headless>=4.8",
    "numpy>=1.24",
    "rembg>=2.0",
    "openai>=1.30",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "aiofiles>=23.0",
    "python-multipart>=0.0.6",
    "cryptography>=42.0",
    "invisible-watermark>=0.2.0",
    "PyWavelets>=1.4",
]

[project.optional-dependencies]
gpu-directml = ["onnxruntime-directml"]
gpu-cuda = ["onnxruntime-gpu"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[project.scripts]
needicons = "needicons.cli:main"
